Shortrocity/requirements.txt
openai
python-dotenv
deepgram-sdk
together
opencv-python
pydub
# ffmpeg-python
